Split up the Renamer module into submodules
