+++ lib/haskell/natural4/src/LS/Interpreter.hs
@@ -112,6 +112,9 @@ import LS.RelationalPredicates
     getBSR,
     partitionExistentials,
   )
+import LS.Renamer.Rules (RnRule)
+import qualified LS.Renamer as Renamer
+import qualified LS.Renamer.Scope as Scope
 import LS.Rule
   ( Rule
       ( DefTypically,
@@ -215,8 +218,6 @@ import LS.XPile.Logging
   )
 import Text.Pretty.Simple (pShowNoColor)
 import Text.Regex.PCRE.Heavy qualified as PCRE
-import qualified LS.Renamer as Renamer
-
 
 -- | This is generated by the Interpreter and handed around to different transpilers.
 --
@@ -243,7 +244,7 @@ data Interpreted = L4I {
   -- eliminated. [TODO].
   , origrules  :: [Rule]
 
-  , renamedRules :: Renamer.RenamerResult [Renamer.RnRule]
+  , renamedRules :: Renamer.RenamerResult [RnRule]
 
   -- | valuepredicates contain the bulk of the top-level decision logic, and can be easily expressed as instance or class methosd.
   , valuePreds :: [ValuePredicate]
@@ -260,7 +261,7 @@ defaultL4I = L4I
   { classtable = CT Map.empty
   , scopetable = Map.empty
   , origrules = mempty
-  , renamedRules = Renamer.RenamerSuccess mempty Renamer.emptyScope
+  , renamedRules = Renamer.RenamerSuccess mempty Scope.emptyScope
   , valuePreds = mempty
   , ruleGraph = Gr.empty
   , ruleGraphErr = mempty
